add: split up NodeRenderer, adds imgui folder
